% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R, R/summary_function.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarise a names data frame}
\usage{
summarize_data(data)

summarize_data(data)
}
\arguments{
\item{data}{A data frame.}

\item{df}{A data.frame}

\item{value_col}{Name of the numeric column to summarise (default tries "finalEstimate")}

\item{group_col}{Name of the grouping column (default tries "cleanName")}
}
\value{
A tibble of column means.
}
\description{
Calculates the mean of all numeric variables in a data frame.
}
\examples{
summarize_data(mtcars)
}
