% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarise numeric columns}
\usage{
summarize_data(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A tibble of column means.
}
\description{
Calculates the mean of all numeric variables in a data frame.
}
\examples{
summarize_data(mtcars)
}
